requirements.txt
dj-database-url==0.3.0
Django==1.8.1
django-postgrespool==0.3.0
gunicorn==19.4.1
psycopg2==2.6.1
SQLAlchemy==1.0.9
whitenoise==2.0.6
requests==2.8.1
pinax_theme_bootstrap==7.2.0
django-user-accounts==1.2.0
django-bootstrap-form==3.2
django-datetime-widget==0.9.3
django-tables2==1.0.4
django-filter==0.11.0